Rename files to camel case notation
