+++ taphome_sdk/__init__.py
@@ -1,13 +1,13 @@
-from .CoverService import CoverService, CoverState
-from .Device import Device
-from .LightService import LightService, LightState
-from .Location import Location
-from .MultiValueSwitchService import MultiValueSwitchService, MultiValueSwitchState
-from .SwitchService import SwitchService, SwitchState
-from .SwitchStates import SwitchStates
-from .TapHomeApiService import TapHomeApiService
-from .TapHomeHttpClientFactory import TapHomeHttpClientFactory
-from .ThermostatService import ThermostatService, ThermostatState
-from .ValueChangeResult import ValueChangeResult
-from .ValueType import ValueType
+from .cover_service import CoverService, CoverState
+from .device import Device
+from .light_service import LightService, LightState
+from .location import Location
+from .multi_value_switch_service import MultiValueSwitchService, MultiValueSwitchState
+from .switch_service import SwitchService, SwitchState
+from .switch_states import SwitchStates
+from .taphome_api_service import TapHomeApiService
+from .taphome_http_client_factory import TapHomeHttpClientFactory
+from .thermostat_service import ThermostatService, ThermostatState
+from .value_change_result import ValueChangeResult
+from .value_type import ValueType
 from .taphome_device_state import TapHomeState
