+++ taphome_sdk/light_service.py
@@ -1,8 +1,8 @@
-from .Device import Device
-from .SwitchStates import SwitchStates
-from .TapHomeApiService import TapHomeApiService
-from .ValueChangeResult import ValueChangeResult
-from .ValueType import ValueType
+from .device import Device
+from .switch_states import SwitchStates
+from .taphome_api_service import TapHomeApiService
+from .value_change_result import ValueChangeResult
+from .value_type import ValueType
 from .taphome_device_state import TapHomeState
