+++ taphome_sdk/switch_service.py
@@ -2,10 +2,10 @@ import logging
 
 from voluptuous.schema_builder import Undefined
 
-from .Device import Device
-from .SwitchStates import SwitchStates
-from .TapHomeApiService import TapHomeApiService
-from .ValueType import ValueType
+from .device import Device
+from .switch_states import SwitchStates
+from .taphome_api_service import TapHomeApiService
+from .value_type import ValueType
 from .taphome_device_state import TapHomeState
 
 _LOGGER = logging.getLogger(__name__)
